chore: consolidate animations.ts file loc
