+++ apps/web/components/base/alert-dialog.tsx
@@ -5,7 +5,7 @@ import * as React from 'react'
 
 import { buttonVariants } from '~/components/base/button'
 import useReducedMotion from '~/hooks/use-reduced-motion'
-import { animations } from '~/lib/animations'
+import { animations } from '~/lib/constants/animations'
 import { cn } from '~/lib/utils'
 
 const AlertDialog = AlertDialogPrimitive.Root
@@ -149,14 +149,14 @@ AlertDialogCancel.displayName = AlertDialogPrimitive.Cancel.displayName
 
 export {
 	AlertDialog,
-	AlertDialogPortal,
-	AlertDialogOverlay,
-	AlertDialogTrigger,
-	AlertDialogContent,
-	AlertDialogHeader,
-	AlertDialogFooter,
-	AlertDialogTitle,
-	AlertDialogDescription,
 	AlertDialogAction,
 	AlertDialogCancel,
+	AlertDialogContent,
+	AlertDialogDescription,
+	AlertDialogFooter,
+	AlertDialogHeader,
+	AlertDialogOverlay,
+	AlertDialogPortal,
+	AlertDialogTitle,
+	AlertDialogTrigger,
 }
