+++ apps/web/components/base/dialog.tsx
@@ -4,7 +4,7 @@ import * as DialogPrimitive from '@radix-ui/react-dialog'
 import { X } from 'lucide-react'
 import * as React from 'react'
 import useReducedMotion from '~/hooks/use-reduced-motion'
-import { animations } from '~/lib/animations'
+import { animations } from '~/lib/constants/animations'
 import { cn } from '~/lib/utils'
 
 const Dialog = DialogPrimitive.Root
@@ -130,13 +130,13 @@ DialogDescription.displayName = DialogPrimitive.Description.displayName
 
 export {
 	Dialog,
-	DialogPortal,
-	DialogOverlay,
-	DialogTrigger,
 	DialogClose,
 	DialogContent,
-	DialogHeader,
-	DialogFooter,
-	DialogTitle,
 	DialogDescription,
+	DialogFooter,
+	DialogHeader,
+	DialogOverlay,
+	DialogPortal,
+	DialogTitle,
+	DialogTrigger,
 }
