+++ apps/web/components/base/sheet.tsx
@@ -5,7 +5,7 @@ import { type VariantProps, cva } from 'class-variance-authority'
 import { X } from 'lucide-react'
 import * as React from 'react'
 import useReducedMotion from '~/hooks/use-reduced-motion'
-import { animations } from '~/lib/animations'
+import { animations } from '~/lib/constants/animations'
 import { cn } from '~/lib/utils'
 
 const Sheet = SheetPrimitive.Root
